Move init_app into flask_init module
We try to keep our __init__ modules as clear as possible and import from
sub modules where possible. With the addition of init_manager in the
next commit the motivation becomes more.
